+++ app/__init__.py
@@ -20,7 +20,7 @@ import datetime, time, re, os, sys, string, json, html
 from functools import wraps
 from flask_login import LoginManager
 
-__version__ = '0.7.0'
+__version__ = '0.8-dev'
 
 # we want to support Python 3 only.
 # Python 2 has too many caveats.